# Conflicts:
#	src/test/java/de/hhu/propra16/avaders/catalogueLoader/test.xml

+++ src/main/java/de/hhu/propra16/avaders/gui/tools/FileTools.java
@@ -1,35 +1,32 @@
-package de.hhu.propra16.avaders.gui;
-
-import de.hhu.propra16.avaders.catalogueLoader.tokenizer.LineReader;
+package de.hhu.propra16.avaders.gui.tools;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-/**
- * Created by Batman140 on 10.07.2016.
- */
 public class FileTools {
-
+	//finished
 	public static void createFile(Path path, String output){
 		if(path == null){
 			System.err.println("Path is null");
 			return;
 		}
-		List<String> outputList = new ArrayList<>();
-		outputList = Arrays.asList(output.split("\\n"));
-		deleteFile(path);
+		List<String> outputList = Arrays.asList(output.split("\\n"));
+		if(Files.exists(path))
+			return;
 		try {
 			Files.createDirectories(path.subpath(0, path.getNameCount()-1));
 			Files.write(path, outputList, StandardOpenOption.CREATE);
+			System.out.println("" + path.toString() + " created!");
 		} catch (IOException e) {
 			System.err.println("Unable to write to File " + path);
 		}
 	}
+
+	//finished
 	private static void deleteFile(Path path){
 		if(!Files.exists(path))
 			return;
@@ -40,9 +37,9 @@ public class FileTools {
 		}
 	}
 
-	//TODO check validity
+	//finished
 	public static String readFile(Path path){
-		List<String> content = null;
+		List<String> content;
 		try {
 			content = Files.readAllLines(path);
 		} catch (IOException e) {
@@ -58,6 +55,18 @@ public class FileTools {
 			output += line + "\n";
 		}
 		return output;
+	}
+
+	public static void updateFile(Path path, String output){
+		if(Files.exists(path))
+			deleteFile(path);
+		List<String> outputList = Arrays.asList(output.split("\\n"));
+		try {
+			Files.write(path, outputList, StandardOpenOption.CREATE);
+			System.out.println("" + path.toString() + " created!");
+		} catch (IOException e) {
+			System.err.println("Unable to update to File " + path);
+		}
 	}
 
 }
